// src/sdk/models/Department.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Department = {
  /**
   * Department ID
   */
  id: string;
  /**
   * Department name
   */
  name: string;
  /**
   * Company ID
   */
  companyId: string;
  /**
   * Number of people in the department
   */
  headcount: number;
  /**
   * Department description
   */
  description?: string;
  /**
   * Whether the department is active
   */
  isActive: boolean;
  /**
   * Creation timestamp
   */
  createdAt: string;
  /**
   * Last update timestamp
   */
  updatedAt: string;
  /**
   * User who last modified the record
   */
  modifiedBy?: string;
  /**
   * Last modification timestamp
   */
  modifiedAt?: string;
  /**
   * Company information
   */
  company?: Record<string, any>;
};